Support explicit array creation in PropertyInvalidateAttribute
