+++ src/MvvmGen.SourceGenerators/Extensions/StringExtensions.cs
@@ -6,9 +6,9 @@
 
 namespace MvvmGen.Extensions
 {
-    public static class StringExtensions
+    internal static class StringExtensions
     {
-        public static string ToCamelCase(this string str)
+        internal static string ToCamelCase(this string str)
         {
             return str.Substring(0, 1).ToLower() + str.Substring(1);
         }
